Finished all DAO tests, passes all tests

+++ server/src/main/java/dataAccess/SQLAuthDAO.java
@@ -6,8 +6,6 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.UUID;
 
 import static java.sql.Statement.RETURN_GENERATED_KEYS;
@@ -16,31 +14,13 @@ import static java.sql.Types.NULL;
 public class SQLAuthDAO implements AuthDAO {
 
     public AuthData addAuth (String username) throws DataAccessException {
+        if (username == "") throw new DataAccessException("Error: unauthorized");
         String token = UUID.randomUUID().toString();
         AuthData data = new AuthData(token, username);
-
         String statement = "INSERT INTO authData (token, username) VALUES (?, ?)";
         executeUpdate(statement, token, username);
 
         return data;
-    }
-    //I just got rid of response exceptions, we'll find out later if we need them. If need, reference petshop.
-    public Collection<AuthData> listAuth() throws DataAccessException {
-        ArrayList<AuthData> authData = new ArrayList<>();
-        try (Connection conn = DatabaseManager.getConnection()) {
-            String statement = "SELECT token, username FROM authData";
-            try (PreparedStatement ps = conn.prepareStatement(statement)) {
-                try (ResultSet rs = ps.executeQuery()) {
-                    while (rs.next()) {
-                        authData.add(makeData(rs));
-                    }
-                }
-            }
-        }
-        catch (DataAccessException | SQLException accessException) {
-            throw new DataAccessException(accessException.getMessage());
-        }
-        return (authData);
     }
 
     public AuthData getAuth(String id) throws DataAccessException {
@@ -66,19 +46,14 @@ public class SQLAuthDAO implements AuthDAO {
     }
 
     public void deleteAuthData(String id)  throws DataAccessException {
+        if (getAuth(id) == null) throw new DataAccessException("authToken DNE");
         String statement = "DELETE FROM authData WHERE token = ?";
         executeUpdate(statement, id);
     }
 
     public void clear()  throws DataAccessException {
-        String statement = "TRUNCATE authData"; //think about using truncate users
+        String statement = "TRUNCATE authData";
         executeUpdate(statement);
-    }
-
-    private AuthData makeData(ResultSet rs) throws SQLException {
-        String token = UUID.randomUUID().toString();
-        String username = rs.getString("username");
-        return new AuthData(token, username);
     }
 
     private void executeUpdate(String statement, Object... params) throws DataAccessException {
